# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface MessageResponse {
  success: boolean;
}

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  image: string | null;
}

export interface RegisterInput {
  /** @minLength 3 */
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name?: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}