# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      description: Creates an account with email/password and starts a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      description: Verifies credentials and starts a session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      description: Destroys the current session and clears the session cookie.
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current authenticated user
      description: Returns the user for the active session, or 401 if not signed in.
      responses:
        "200":
          description: The current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/google:
    get:
      operationId: startGoogleAuth
      tags: [auth]
      summary: Start Google OAuth
      description: >
        Browser-redirect endpoint. Redirects the user to Google's consent
        screen (authorization-code flow with PKCE). Navigate to this URL
        directly via the browser; it is not a JSON/fetch endpoint.
      responses:
        "302":
          description: Redirect to Google's OAuth consent screen
  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [auth]
      summary: Google OAuth callback
      description: >
        Browser-redirect endpoint. Google redirects back here with an
        authorization code; the server exchanges it, creates or links the
        user, starts a session, and redirects into the web app.
      responses:
        "302":
          description: Redirect back into the web app
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
      required:
        - id
        - email
        - name
        - image
    RegisterInput:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 3
        password:
          type: string
          minLength: 8
        name:
          type: string
          minLength: 1
      required:
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required:
        - email
        - password